Backend/requirements.txt
Django==2.1.3
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
oauthlib==2.0.3
PyJWT==1.6.4